requirements.txt
numpy
requests
prettytable
rich
torch
torchvision
torchaudio
pandas
google-cloud-bigquery==3.10.0
google-cloud-storage==2.8.0